Do not pass all props to the wrapper div
React shows this warning for invalid props on regular HTML elements.

```
Warning: Unknown prop `onClickOutside` on <div> tag. Remove this prop from the element. For details, see https://fb.me/react-unknown-prop
```

We need to avoid using `{...this.props}` on those elements, and generally it doesn't work with PropTypes so not a very good idea.